Add gradle build to project
Also rename to burst-service for the core codebase to allow extra modules to be added
Move code into gradle/normal format layout src/main/java
Move schema into main/resources
Move test xmls into test/resources
Move tester classes into utils/test
