+++ burst-service/build.gradle
@@ -1,0 +1,1 @@
+apply plugin: 'maven'
