+++ burst-service/src/main/java/ox/softeng/burst/domain/Severity.java
