+++ burst-service/src/main/java/ox/softeng/burst/domain/User.java
