+++ burst-service/src/main/java/ox/softeng/burst/services/Main.java
@@ -1,8 +1,8 @@
 package ox.softeng.burst.services;
 
+import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.Executors;
 
 public class Main {
