+++ burst-service/src/main/java/ox/softeng/burst/services/RabbitService.java
