+++ burst-service/src/main/java/ox/softeng/burst/util/HibernateUtil.java
