+++ burst-service/src/main/java/ox/softeng/burst/util/test/TestSendMessage.java
@@ -1,4 +1,4 @@
-package ox.softeng.burst;
+package ox.softeng.burst.util.test;
 
 import com.rabbitmq.client.Connection;
 import com.rabbitmq.client.ConnectionFactory;
