+++ burst-service/src/main/resources/schema/messageSchema.xsd
