+++ burst-service/src/test/resources/message/instance1.xml
