+++ burst-service/src/test/resources/message/instance2.xml
