+++ burst-service/src/test/resources/message/instance3.xml
