+++ burst-service/src/test/resources/message/instance4.xml
