+++ burst-service/src/test/resources/message/instance6.xml
