+++ burst-service/src/test/resources/message/instance7.xml
