+++ burst-service/src/test/resources/message/instance8.xml
