+++ dependencies.gradle
@@ -1,0 +1,33 @@
+/**
+ * Only dependencies which are required across ALL projects should be in here
+ */
+dependencies {
+
+    /**
+     * Logging
+     */
+    // Use slf4j logging API and force all log4j over slf4j
+    compile group: 'org.slf4j', name: 'slf4j-api', version: slf4jVersion
+    compile group: 'org.slf4j', name: 'log4j-over-slf4j', version: slf4jVersion, transitive: false
+    compile group: 'org.slf4j', name: 'jul-to-slf4j', version: slf4jVersion
+    compile group: 'org.slf4j', name: 'jcl-over-slf4j', version: slf4jVersion
+
+    runtime group: 'ch.qos.logback', name: 'logback-classic', version: logbackVersion, transitive: false
+    runtime group: 'ch.qos.logback', name: 'logback-core', version: logbackVersion
+
+    /**
+     * Testing
+     */
+    testCompile group: 'org.spockframework', name: 'spock-core', version: '1.0-groovy-2.4'
+    testCompile group: 'junit', name: 'junit', version: '4.12'
+}
+
+configurations.all {
+    resolutionStrategy.eachDependency {DependencyResolveDetails details ->
+        if (details.requested.group == 'org.postgresql' && details.requested.name == 'postgresql'
+                && details.requested.version == '9.4.1207.jre7') {
+            //prefer different version which contains some necessary fixes
+            details.useVersion '9.4.1207'
+        }
+    }
+}
