+++ gradle.properties
@@ -1,0 +1,26 @@
+version=1.0
+group=ox.softeng.burst
+
+org.gradle.daemon=true
+gradleVersion=2.11
+
+sourceCompatibility=1.8
+targetCompatibility=1.8
+
+slf4jVersion=1.7.16
+logbackVersion=1.1.3
+
+hibernateVersion=5.0.6.Final
+hibernateCommonsVersion=5.0.1.Final
+hibernateValidatorVersion=5.2.2.Final
+postgresJdbcVersion=9.4.1207
+
+rabbitmqVersion=3.6.0
+
+grailsVersion=3.1.1
+gormVersion=5.0.1
+groovyVersion=2.4.5
+
+gsonVersion=2.3.1
+commonsCodecVersion=1.10
+springframeworkVersion=4.2.4.RELEASE
