+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Feb 18 10:42:14 GMT 2016
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-2.11-bin.zip
